feat: Call huggingface API for recipie response
Implement a netlify serverless function for requesting a recipie using
added ingredients, from the huggingface AI API. This requires an
environment variable of HF_API_KEY, corresponding to the API key for
hugging face.

In future, it would be good to implement a toggle for development, to
abstract / isolate the netlify function.